+++ lib/DxilValidation/DxilValidation.cpp
@@ -35,6 +35,7 @@
 #include "llvm/Analysis/ValueTracking.h"
 #include "llvm/Bitcode/ReaderWriter.h"
 #include "llvm/IR/Constants.h"
+#include "llvm/IR/DerivedTypes.h"
 #include "llvm/IR/DiagnosticInfo.h"
 #include "llvm/IR/DiagnosticPrinter.h"
 #include "llvm/IR/Dominators.h"
@@ -995,6 +996,58 @@ static void ValidateLinAlgOpParameters(CallInst *CI,
   }
 }
 
+static void ValidateLinAlgComponentType(CallInst *CI, DXIL::ComponentType CT,
+                                        ValidationContext &ValCtx) {
+  switch (CT) {
+  case DXIL::ComponentType::I8:
+  case DXIL::ComponentType::I16:
+  case DXIL::ComponentType::I32:
+  case DXIL::ComponentType::I64:
+  case DXIL::ComponentType::U8:
+  case DXIL::ComponentType::U16:
+  case DXIL::ComponentType::U32:
+  case DXIL::ComponentType::U64:
+  case DXIL::ComponentType::F8_E4M3FN:
+  case DXIL::ComponentType::F8_E5M2:
+  case DXIL::ComponentType::F16:
+  case DXIL::ComponentType::F32:
+  case DXIL::ComponentType::F64:
+  case DXIL::ComponentType::BFloat16:
+    break;
+  default:
+    ValCtx.EmitInstrFormatError(CI,
+                                ValidationRule::InstrLinAlgIllegalComponentType,
+                                {ComponentTypeToString(CT)});
+    break;
+  }
+}
+
+static unsigned ComponentTypeElementsPerScalar(DXIL::ComponentType CT) {
+  switch (CT) {
+  case DXIL::ComponentType::I16:
+  case DXIL::ComponentType::I32:
+  case DXIL::ComponentType::I64:
+  case DXIL::ComponentType::U16:
+  case DXIL::ComponentType::U32:
+  case DXIL::ComponentType::U64:
+  case DXIL::ComponentType::F16:
+  case DXIL::ComponentType::F32:
+  case DXIL::ComponentType::F64:
+    return 1;
+  case DXIL::ComponentType::BFloat16:
+    return 2;
+  case DXIL::ComponentType::I8:
+  case DXIL::ComponentType::U8:
+  case DXIL::ComponentType::F8_E4M3FN:
+  case DXIL::ComponentType::F8_E5M2:
+    return 4;
+  // All other ComponentTypes are illegal to use in LinAlg Matrix. Their usage
+  // is detected and reported in other parts on the validator
+  default:
+    return 4;
+  }
+}
+
 static void ValidateLinAlgOpReturnMatrix(CallInst *CI,
                                          ValidationContext &ValCtx) {
   Type *Ty = CI->getType();
@@ -1025,28 +1078,7 @@ static void ValidateLinAlgOpReturnMatrix(CallInst *CI,
           {std::to_string(K), std::to_string(MinK), std::to_string(MaxK)});
   }
 
-  // Validate the ComponentType is allowed
-  switch (LATT.Type) {
-  case DXIL::ComponentType::I8:
-  case DXIL::ComponentType::I16:
-  case DXIL::ComponentType::I32:
-  case DXIL::ComponentType::I64:
-  case DXIL::ComponentType::U8:
-  case DXIL::ComponentType::U16:
-  case DXIL::ComponentType::U32:
-  case DXIL::ComponentType::U64:
-  case DXIL::ComponentType::F8_E4M3FN:
-  case DXIL::ComponentType::F8_E5M2:
-  case DXIL::ComponentType::F16:
-  case DXIL::ComponentType::F32:
-  case DXIL::ComponentType::F64:
-    break;
-  default:
-    ValCtx.EmitInstrFormatError(CI,
-                                ValidationRule::InstrLinAlgIllegalComponentType,
-                                {ComponentTypeToString(LATT.Type)});
-    break;
-  }
+  ValidateLinAlgComponentType(CI, LATT.Type, ValCtx);
 }
 
 static void ValidateLinAlgMatrixLength(CallInst *CI,
@@ -1074,19 +1106,195 @@ static void ValidateLinAlgMatrixStoreToMemory(CallInst *CI,
   ValidateLinAlgOpParameters(CI, ValCtx);
 }
 
-static void ValidateLinAlgMatVecMul(CallInst *CI, ValidationContext &ValCtx) {
+static void ValidateLinAlgMatVecMul(CallInst *CI, ValidationContext &ValCtx,
+                                    const char *OpName = "LinAlgMatVecMul") {
   ValidateLinAlgOpParameters(CI, ValCtx);
+
+  DxilInst_LinAlgMatVecMul Op(CI);
+
+  VectorType *OutputVecTy = cast<VectorType>(CI->getType());
+  Type *MatTy = Op.get_matrix()->getType();
+  ConstantInt *IsSignedCI = dyn_cast<ConstantInt>(Op.get_isOutputSigned());
+  VectorType *InputVecTy = cast<VectorType>(Op.get_inputVector()->getType());
+  ConstantInt *InputInterpCI = dyn_cast<ConstantInt>(Op.get_interpretation());
+
+  assert(dxilutil::IsHLSLLinAlgMatrixType(MatTy) && "Must be LinAlg type");
+
+  auto MatIt = ValCtx.LinAlgTargetTypeMap.find(MatTy);
+  if (MatIt == ValCtx.LinAlgTargetTypeMap.end())
+    return;
+  LinAlgTargetType MatLATT = MatIt->second;
+
+  // Mat must be A matrix of Thread scope
+  if (MatLATT.Scope != DXIL::MatrixScope::Thread)
+    ValCtx.EmitInstrFormatError(CI,
+                                ValidationRule::InstrLinAlgMatrixScopeMismatch,
+                                {MatrixScopeToString(MatLATT.Scope), "Thread"});
+  if (MatLATT.Use != DXIL::MatrixUse::A)
+    ValCtx.EmitInstrFormatError(CI,
+                                ValidationRule::InstrLinAlgMatrixUseMismatch,
+                                {MatrixUseToString(MatLATT.Use), "A"});
+
+  // Input Interp must be a immarg of allowed ComponentType
+  DXIL::ComponentType Interp = DXIL::ComponentType::Invalid;
+  if (InputInterpCI) {
+    Interp = static_cast<DXIL::ComponentType>(InputInterpCI->getLimitedValue());
+    ValidateLinAlgComponentType(CI, Interp, ValCtx);
+  } else
+    ValCtx.EmitInstrFormatError(CI, ValidationRule::InstrOpConst,
+                                {"InputInterp", OpName});
+
+  // InputVec's length must match the K dim of input matrix after accounting
+  // for multiple elements packed into a single scalar. The packed elements may
+  // not fully saturate the final vector element but it must still be included.
+  // K is always the N of the matrix since its ensured to be an A Matrix.
+  unsigned ElementsPerScalar = ComponentTypeElementsPerScalar(Interp);
+  unsigned ExpectedVecK =
+      (MatLATT.N + ElementsPerScalar - 1) / ElementsPerScalar;
+
+  if (ExpectedVecK != InputVecTy->getNumElements())
+    ValCtx.EmitInstrFormatError(
+        CI, ValidationRule::InstrLinAlgMatrixDimKVecKMismatch,
+        {"Input", std::to_string(InputVecTy->getNumElements()),
+         std::to_string(ExpectedVecK), std::to_string(MatLATT.N),
+         ComponentTypeToString(Interp)});
+
+  // OutputVec length must match M dim
+  if (MatLATT.M != OutputVecTy->getNumElements())
+    ValCtx.EmitInstrFormatError(
+        CI, ValidationRule::InstrLinAlgMatrixDimVectorMismatch,
+        {"Output", std::to_string(OutputVecTy->getNumElements()), "M",
+         std::to_string(MatLATT.M)});
+
+  // Sign bit must be immarg and must be true if output vec is a
+  // native floating point type
+  if (IsSignedCI) {
+    bool IsSigned = IsSignedCI->isOne();
+    if (OutputVecTy->getElementType()->isFloatingPointTy() && !IsSigned)
+      ValCtx.EmitInstrFormatError(
+          CI, ValidationRule::InstrLinAlgMatrixUnsignedFloatTypeNotAllowed,
+          {TypeToString(OutputVecTy->getElementType())});
+  } else
+    ValCtx.EmitInstrFormatError(CI, ValidationRule::InstrOpConst,
+                                {"IsSigned", OpName});
 }
 
 static void ValidateLinAlgMatVecMulAdd(CallInst *CI,
                                        ValidationContext &ValCtx) {
-  ValidateLinAlgOpParameters(CI, ValCtx);
+  // All the rules from LinAlgMatVecMul apply
+  ValidateLinAlgMatVecMul(CI, ValCtx, "LinAlgMatVecMulAdd");
+
+  DxilInst_LinAlgMatVecMulAdd Op(CI);
+
+  VectorType *OutputVecTy = cast<VectorType>(CI->getType());
+  Type *MatTy = Op.get_matrix()->getType();
+  auto MatIt = ValCtx.LinAlgTargetTypeMap.find(MatTy);
+  if (MatIt == ValCtx.LinAlgTargetTypeMap.end())
+    return;
+  LinAlgTargetType MatLATT = MatIt->second;
+  VectorType *BiasVecTy = cast<VectorType>(Op.get_biasVector()->getType());
+
+  // BiasVec length must match M dim
+  if (MatLATT.M != BiasVecTy->getNumElements())
+    ValCtx.EmitInstrFormatError(
+        CI, ValidationRule::InstrLinAlgMatrixDimVectorMismatch,
+        {"Bias", std::to_string(BiasVecTy->getNumElements()), "M",
+         std::to_string(MatLATT.M)});
+
+  // Bias element type must match output element type
+  if (BiasVecTy->getElementType() != OutputVecTy->getElementType())
+    ValCtx.EmitInstrFormatError(
+        CI, ValidationRule::InstrLinAlgMatrixOutputBiasVecMismatch,
+        {TypeToString(OutputVecTy->getElementType()),
+         TypeToString(BiasVecTy->getElementType())});
 }
 
 static void
 ValidateLinAlgMatrixAccumulateToDescriptor(CallInst *CI,
                                            ValidationContext &ValCtx) {
   ValidateLinAlgOpParameters(CI, ValCtx);
+
+  DxilInst_LinAlgMatrixAccumulateToDescriptor Op(CI);
+  Type *MatTy = Op.get_matrix()->getType();
+
+  assert(dxilutil::IsHLSLLinAlgMatrixType(MatTy) && "Must be LinAlg type");
+  auto MatIt = ValCtx.LinAlgTargetTypeMap.find(MatTy);
+  if (MatIt == ValCtx.LinAlgTargetTypeMap.end())
+    return;
+  LinAlgTargetType MatLATT = MatIt->second;
+
+  ConstantInt *LayoutCI = dyn_cast<ConstantInt>(Op.get_layout());
+  if (!LayoutCI) {
+    ValCtx.EmitInstrFormatError(
+        CI, ValidationRule::InstrOpConst,
+        {"Layout", "LinAlgMatrixAccumulateToDescriptor"});
+    return;
+  }
+  auto Layout = static_cast<DXIL::MatrixLayout>(LayoutCI->getLimitedValue());
+  bool LayoutIsRowColMajor = (Layout == DXIL::MatrixLayout::RowMajor ||
+                              Layout == DXIL::MatrixLayout::ColumnMajor);
+
+  // Thread Matrix must have layout OuterProductOptimal*
+  if (MatLATT.Scope == DXIL::MatrixScope::Thread &&
+      (Layout != DXIL::MatrixLayout::OuterProductOptimal &&
+       Layout != DXIL::MatrixLayout::OuterProductOptimalTranspose))
+    ValCtx.EmitInstrFormatError(
+        CI, ValidationRule::InstrLinAlgMatrixScopeReqLayout2,
+        {MatrixScopeToString(MatLATT.Scope), "OuterProductOptimal",
+         "OuterProductOptimalTranspose"});
+
+  // Wave/ThreadGroup matrix must have layout RowMajor/ColMajor
+  if (MatLATT.Scope != DXIL::MatrixScope::Thread && !LayoutIsRowColMajor)
+    ValCtx.EmitInstrFormatError(
+        CI, ValidationRule::InstrLinAlgMatrixScopeReqLayout2,
+        {MatrixScopeToString(MatLATT.Scope), "RowMajor", "ColumnMajor"});
+
+  // Stride must be an imm 0 if layout is not Row/Col Major
+  if (!LayoutIsRowColMajor) {
+    ConstantInt *StrideCI = dyn_cast<ConstantInt>(Op.get_stride());
+    if (StrideCI) {
+      if (!StrideCI->isZero())
+        ValCtx.EmitInstrFormatError(
+            CI, ValidationRule::InstrLinAlgMatrixLayoutReqStride,
+            {MatrixLayoutToString(Layout)});
+    } else
+      ValCtx.EmitInstrFormatError(
+          CI, ValidationRule::InstrOpConst,
+          {"Stride", "LinAlgMatrixAccumulateToDescriptor"});
+  }
+
+  // Matrix must have Accumulator use
+  if (MatLATT.Use != DXIL::MatrixUse::Accumulator)
+    ValCtx.EmitInstrFormatError(
+        CI, ValidationRule::InstrLinAlgMatrixUseMismatch,
+        {MatrixUseToString(MatLATT.Use), "Accumulator"});
+
+  // handle must be a UAV Raw buffer (RWByteAddressBuffer)
+  DXIL::ComponentType ResCompTy;
+  DXIL::ResourceClass ResClass;
+  DXIL::ResourceKind ResKind =
+      GetResourceKindAndCompTy(Op.get_handle(), ResCompTy, ResClass, ValCtx);
+  if (ResClass != DXIL::ResourceClass::UAV ||
+      ResKind != DXIL::ResourceKind::RawBuffer)
+    ValCtx.EmitInstrFormatError(CI,
+                                ValidationRule::InstrLinAlgMatrixRequiresRWBAB,
+                                {"LinAlgMatrixAccumulateToDescriptor"});
+
+  // Align must be an immediate constant that is a multiple of 128 greater than
+  // 0
+  ConstantInt *AlignCI = dyn_cast<ConstantInt>(Op.get_align());
+  if (AlignCI) {
+    unsigned Align = AlignCI->getLimitedValue();
+    if (Align == 0)
+      ValCtx.EmitInstrFormatError(CI, ValidationRule::InstrParamMinimumValue,
+                                  {"Align", "0", std::to_string(Align)});
+    if (Align % 128 != 0)
+      ValCtx.EmitInstrFormatError(CI, ValidationRule::InstrParamMultiple,
+                                  {"Align", "128", std::to_string(Align)});
+  } else
+    ValCtx.EmitInstrFormatError(
+        CI, ValidationRule::InstrOpConst,
+        {"Align", "LinAlgMatrixAccumulateToDescriptor"});
 }
 
 static void ValidateLinAlgMatrixAccumulateToMemory(CallInst *CI,
@@ -1144,16 +1352,16 @@ static void ValidateLinAlgMatrixLoadFromDescriptor(CallInst *CI,
   ValidateLinAlgOpReturnMatrix(CI, ValCtx);
   ValidateLinAlgOpParameters(CI, ValCtx);
 
+  DxilInst_LinAlgMatrixLoadFromDescriptor Op(CI);
   Type *RetMatTy = CI->getType();
+
   assert(dxilutil::IsHLSLLinAlgMatrixType(RetMatTy) && "Must be LinAlg type");
   auto RetIt = ValCtx.LinAlgTargetTypeMap.find(RetMatTy);
   if (RetIt == ValCtx.LinAlgTargetTypeMap.end())
     return;
   LinAlgTargetType RetLATT = RetIt->second;
 
-  Value *StrideOp = CI->getArgOperand(3);
-  Value *LayoutOp = CI->getArgOperand(4);
-  ConstantInt *LayoutCI = dyn_cast<ConstantInt>(LayoutOp);
+  ConstantInt *LayoutCI = dyn_cast<ConstantInt>(Op.get_layout());
   if (!LayoutCI) {
     ValCtx.EmitInstrFormatError(CI, ValidationRule::InstrOpConst,
                                 {"Layout", "LinAlgMatrixLoadFromDescriptor"});
@@ -1172,7 +1380,7 @@ static void ValidateLinAlgMatrixLoadFromDescriptor(CallInst *CI,
 
   // Stride must be an imm 0 if Layout is not Row/Col Major
   if (!LayoutIsRowColMajor) {
-    ConstantInt *StrideCI = dyn_cast<ConstantInt>(StrideOp);
+    ConstantInt *StrideCI = dyn_cast<ConstantInt>(Op.get_stride());
     if (StrideCI) {
       if (!StrideCI->isZero())
         ValCtx.EmitInstrFormatError(
@@ -1181,6 +1389,32 @@ static void ValidateLinAlgMatrixLoadFromDescriptor(CallInst *CI,
     } else
       ValCtx.EmitInstrFormatError(CI, ValidationRule::InstrOpConst,
                                   {"Stride", "LinAlgMatrixLoadFromDescriptor"});
+  }
+
+  // Align must be an immediate constant that is a multiple of 128
+  ConstantInt *AlignCI = dyn_cast<ConstantInt>(Op.get_align());
+  if (AlignCI) {
+    unsigned Align = AlignCI->getLimitedValue();
+    if (Align == 0)
+      ValCtx.EmitInstrFormatError(CI, ValidationRule::InstrParamMinimumValue,
+                                  {"Align", "0", std::to_string(Align)});
+    if (Align % 128 != 0)
+      ValCtx.EmitInstrFormatError(CI, ValidationRule::InstrParamMultiple,
+                                  {"Align", "128", std::to_string(Align)});
+  } else
+    ValCtx.EmitInstrFormatError(CI, ValidationRule::InstrOpConst,
+                                {"Align", "LinAlgMatrixLoadFromDescriptor"});
+
+  // Thread matrix may only load from SRV ByteAddressBuffer
+  if (RetLATT.Scope == DXIL::MatrixScope::Thread) {
+    DXIL::ComponentType ResCompTy;
+    DXIL::ResourceClass ResClass;
+    DXIL::ResourceKind ResKind =
+        GetResourceKindAndCompTy(Op.get_handle(), ResCompTy, ResClass, ValCtx);
+    if (ResClass != DXIL::ResourceClass::SRV ||
+        ResKind != DXIL::ResourceKind::RawBuffer)
+      ValCtx.EmitInstrError(
+          CI, ValidationRule::InstrLinAlgMatrixLoadThreadRequiresBAB);
   }
 }